requirements.txt
django
djangorestframework
djangorestframework-simplejwt
urllib3